pyproject.toml
[tool.poetry]
name = "message"
version = "0.0.1"
description = "a general message notification service"
authors = ["Stephen Ling <lingfromsh@163.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
fastapi = "^0.105.0"
pydantic = {extras = ["email"], version = "^2.5.3"}
dependency-injector = "^4.41.0"
redis = "^5.0.1"
strawberry-graphql = { extras = ["fastapi"], version = "^0.216.1" }
tortoise-orm = { extras = ["accel"], version = "^0.20.0" }
uvicorn = {extras = ["standard"], version = "^0.25.0"}
aiofile = "^3.8.8"
trio = "^0.23.2"
pydantic-settings = "^2.1.0"
toml = "^0.10.2"
python-ulid = "^2.2.0"
psutil = "^5.9.7"
rtoml = "^0.10.0"
jsonschema = {extras = ["format"], version = "^4.20.0"}
pydantic-extra-types = "^2.2.0"
phonenumbers = "^8.13.27"
gunicorn = "^21.2.0"
websockets = "^12.0"
backoff = "^2.2.1"
sqlalchemy = {extras = ["postgres"], version = "^2.0.23"}
greenlet = "^3.0.3"
psycopg2 = "^2.9.9"
uvloop = "^0.19.0"
apscheduler = "^3.10.4"
stringcase = "^1.2.0"
blinker = "^1.7.0"
aiomultiprocess = "^0.9.0"
pendulum = "^3.0.0"
aio-pika = "^9.3.1"
asyncpg = "^0.29.0"
loky = "^3.4.1"
taskiq-fastapi = "^0.3.1"
taskiq-redis = "^0.5.5"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "explicit"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
gql = "^3.4.1"
assertpy = "^1.1"


[tool.poetry.group.dev.dependencies]
memray = "^1.11.0"


[tool.poetry.group.prod.dependencies]
supervisor = "^4.2.5"

[tool.isort]
profile = "black"
src_paths = ["src"]
force_single_line = true
import_heading_future = "Future Library"
import_heading_stdlib = "Standard Library"
import_heading_firstparty = "First Library"
import_heading_thirdparty = "Third Party Library"
import_heading_localfolder = "Local Folder"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
